moved RelationshipsHandlerExtensions to a separate file
