+++ ECampus/src/ECampus.Shared/Interfaces/Data/DataServices/IRelationshipsHandler.cs
@@ -13,17 +13,5 @@ public interface IRelationshipsHandler<in TLeftTable, out TRightTable, out TRela
     PropertyInfo RightTableId { get; }
     PropertyInfo LeftTableId { get; }
     PropertyInfo RelatedModels { get; }
-}
-
-public static class RelationshipsHandlerExtensions
-{
-    public static TRelation CreateRelationModel<TRelation, TLeftTable, TRightTable>(
-        this IRelationshipsHandler<TLeftTable, TRightTable, TRelation> handler, int leftTableId, int rightTableId)
-        where TRelation : class, new() where TRightTable : IModel where TLeftTable : IModel
-    {
-        var result = new TRelation();
-        handler.RightTableId.SetMethod?.Invoke(result, new object[] { rightTableId });
-        handler.LeftTableId.SetMethod?.Invoke(result, new object[] { leftTableId });
-        return result;
-    }
+    PropertyInfo RelationModels { get; }
 }
